requirements.txt
geomloss==0.2.6
jedi==0.19.1
matplotlib==3.10.1
netron==8.2.2
numba==0.60.0
numpy==2.2.5
open3d==0.19.0
pandas==2.2.3
peft==0.14.0
plyfile==1.1
PyYAML==6.0.1
PyYAML==6.0.2
scikit_learn==1.6.1
scipy==1.15.2
torch==2.6.0
torchsummary==1.5.1
torchviz==0.0.3
tqdm==4.66.5
transformers==4.49.0
triton==3.2.0
